requirements.txt
paho-mqtt==1.6.1
numpy==1.23.5
boto3==1.24.0